[WIP] Add insert, update and generate html features
With this version, it is possible to create a product, define its
attributes and save it. In the backend, variants will be generated and
inserted into the db. When the product is opened for edition in the
frontend, variants will be displayed in a rough html tab. Fields such
as price, stock and sky can be modified, and once the product is saved,
such modification are update into the db.

This version does not include javascript login in the frontend to
enable the user to add variants and edit them in the frontend before
having inserted them into the db.