Add display sku variants in list of products in orders details
The tricky part of this hook is to find the link between the
order_product_id and the variant_id. Although it works now, a drawback
is that this task must be performed each time we list the products of
an order. We can optimise this by checking if we already have the link.
However, a better place to do this would be after inserting all
products of an order into the db from the shopping cart. It is there
that we should create all links between order_product_id and variant_id
for all products, if possible. This would avoid to rewrite a similar
algorithm is all parts where we could need the link between the
order_product_id and the variant_id.